backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-openai==0.0.2

# Vector Database
chromadb==0.4.18

# Document processing
unstructured==0.11.8
python-docx==1.1.0
PyMuPDF==1.23.8
pypdf==3.17.4

# OpenAI integration
openai==1.6.1

# Environment and utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.2
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Socket.IO for real-time chat
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Additional utilities
Werkzeug==3.0.1
gunicorn==21.2.0